Rename Survey Paper_Jay Mehta.pdf to Ranking and Retrieval of Information from Distributed Databases
Abstract-This paper focuses on various information
retrieval models and common problems faced during the
retrieval of documents from heterogeneous distributed
databases. The search process mainly focuses on
optimizing for resources within the architecture and
discusses several search algorithms and query
scheduling techniques and compares them with the
monolithic systems in-use. We then discuss the efficiency
of these systems and how well can it support data-
intensive applications in practice.